# use IndexByte replace Split to improve performance