benchmark cmp result:
```
benchmark                   old ns/op     new ns/op     delta
BenchmarkParseAccept-12     362           176           -51.38%

benchmark                   old allocs     new allocs     delta
BenchmarkParseAccept-12     6              2              -66.67%

benchmark                   old bytes     new bytes     delta
BenchmarkParseAccept-12     224           128           -42.86%
```